AFID-141 made authentication a generic function ready for multiple uses.
