+++ app/uk/gov/hmrc/taxhistory/controllers/EmploymentHistoryController.scala
@@ -16,64 +16,25 @@
 
 package uk.gov.hmrc.taxhistory.controllers
 
-import play.api.Logger
-import play.api.mvc.Action
+import play.api.mvc.{Action, Result}
 import uk.gov.hmrc.auth.core._
-import uk.gov.hmrc.play.microservice.controller.BaseController
+import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.taxhistory.TaxHistoryAuthConnector
-import uk.gov.hmrc.taxhistory.model.auth.AfiAuth._
 import uk.gov.hmrc.taxhistory.services.EmploymentHistoryService
-import uk.gov.hmrc.auth.core.retrieve.~
 
 import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-import uk.gov.hmrc.http.HeaderCarrier
 
-
-trait EmploymentHistoryController extends BaseController with AuthorisedFunctions {
+trait EmploymentHistoryController extends AuthController {
 
   def employmentHistoryService: EmploymentHistoryService = EmploymentHistoryService
 
-  private def isAgent(group: AffinityGroup): Boolean = group.toString.contains("Agent")
-
-  private def extractArn(enrolls: Set[Enrolment]): Option[String] = {
-    Logger.info("enrolls=" + enrolls)
-
-    def extractArn(enrollmentIndenitifiers: Seq[EnrolmentIdentifier]) = enrollmentIndenitifiers.find(_.key == "AgentReferenceNumber").map(_.value)
-
-    val arn: Option[String] = enrolls.find(_.key equals "HMRC-AS-AGENT").flatMap(agentEnrollment ⇒ extractArn(agentEnrollment.identifiers))
-    Logger.info("arn=" + arn)
-    arn
-  }
-
   def getEmploymentHistory(nino: String, taxYear: Int) = Action.async {
-    implicit request =>
-      authorised(AgentEnrolmentForPAYE.withIdentifier("MTDITID", nino) and AuthProviderAgents).retrieve(affinityGroupAllEnrolls) {
-
-        case Some(affinityG) ~ allEnrols⇒
-
-          Logger.info("allEnrols " + allEnrols)
-          Logger.info("affinityGroup " + affinityG)
-          (isAgent(affinityG), extractArn(allEnrols.enrolments)) match {
-            case (true, Some(arn)) => getTaxHistory(nino, taxYear)
-            case _ =>Future.successful(Unauthorized("Not Authorised"))
-          }
-        case _ => {
-          Logger.debug("failed to retrieve")
-          Future.successful(Unauthorized("Not Authorised"))
-        }
-      }.recoverWith {
-        case i: InsufficientEnrolments =>
-          Logger.error("Error thrown :" + i.getMessage)
-          Future.successful(Unauthorized("Not Authorised"))
-        case e =>
-          Logger.error("Error thrown :" + e.getMessage)
-          Future.successful(InternalServerError(e.getMessage))
-      }
-
+    implicit request => {
+      authorisedRelationship(nino, getTaxHistory(nino, taxYear))
+    }
   }
 
-  private def getTaxHistory(nino: String, taxYear: Int)(implicit hc:HeaderCarrier) = {
+  private def getTaxHistory(nino: String,taxYear: Int)(implicit hc:HeaderCarrier) = {
     employmentHistoryService.getEmploymentHistory(nino, taxYear) map {
       response =>
         response.status match {
